Move the 'id', 'rcu' and 'vec_map' modules under 'util'.
