+++ src/decoder.rs
@@ -8,9 +8,11 @@ use std::sync::Arc;
 use anyhow::{Context, Error, bail};
 
 use crate::capture::prelude::*;
-use crate::rcu::SingleWriterRcu;
 use crate::usb::{self, prelude::*, validate_packet};
-use crate::vec_map::VecMap;
+use crate::util::{
+    rcu::SingleWriterRcu,
+    vec_map::VecMap,
+};
 
 struct EndpointData {
     device_id: DeviceId,
